Reorganize model code to minimize file changes
