+++ infrastructure/api/api_handler/models/noop_job.py
@@ -2,11 +2,16 @@ from typing import Literal, Optional
 
 from pydantic import BaseModel, Extra
 
-from .job import BaseJob, JobType
+from .job import BaseJob, JobType, BaseJobInput
 
 
 class NoOpJobInputs(BaseModel):
     pass
+
+
+class CreateNoOpJobInput(BaseJobInput):
+    job_type: Literal[JobType.NO_OP] = JobType.NO_OP
+    inputs: NoOpJobInputs
 
 
 class NoOpJobOutputs(BaseModel):
